requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.4.0
requests>=2.32.3
httpx>=0.27.0
google-generativeai>=0.7.2
jinja2>=3.1.0
python-dotenv>=1.0.1
moviepy==1.0.3
imageio-ffmpeg>=0.4.7
numpy>=1.24.0
pyngrok>=7.0.0
playwright==1.48.0
playwright-stealth>=1.0.6

# ML Dependencies for Local Panel Detection
transformers==4.49.0
torch
torchvision
accelerate
einops
timm>=0.9.16
pytorch_metric_learning
shapelygroq
